all: Add AS_923 band definition
