+++ pkg/band/all.go
@@ -19,7 +19,15 @@ import (
 )
 
 // All contains all the bands available.
-var All = map[string]map[ttnpb.PHYVersion]Band{}
+var All = map[string]map[ttnpb.PHYVersion]Band{
+	AS_923: {
+		ttnpb.RP001_V1_0_2:       AS_923_RP1_v1_0_2,
+		ttnpb.RP001_V1_0_2_REV_B: AS_923_RP1_v1_0_2_RevB,
+		ttnpb.RP001_V1_0_3_REV_A: AS_923_RP1_v1_0_3_RevA,
+		ttnpb.RP001_V1_1_REV_A:   AS_923_RP1_v1_1_RevA,
+		ttnpb.RP001_V1_1_REV_B:   AS_923_RP1_v1_1_RevB,
+	},
+}
 
 // Get returns the band if it was found, and returns an error otherwise.
 func Get(id string, version ttnpb.PHYVersion) (Band, error) {
